fix(dtype): explicit dtype on all tensor creation — M2378
Megatron M2378: missing dtype= in torch.tensor() creation causes silent
float32 upcast when tensor interacts with bf16 parameters.

Audit and add explicit dtype= to every torch.tensor/zeros/empty/ones
call in distrib_optimizer, finalize_model_grads, param_and_grad_buffer.
Float32 where intentional (norm accum, loss scale) is documented.

DES-LOC: silent float32 upcasts hit A6000 (310 TFLOPS bf16) 3x harder
than H100 (989 TFLOPS bf16), creating asymmetric compute load.

+++ deepspeed/core/distributed/param_and_grad_buffer.py
@@ -1264,7 +1264,7 @@ class ParamAndGradBuffer:
                 )
                 # Warmup NCCL comm buffers.
                 torch.distributed.barrier()
-                tmp = torch.zeros([1], device="cuda")
+                tmp = torch.zeros([1], dtype=torch.float32, device="cuda")  # From Megatron M2378: explicit dtype prevents silent float32 upcast
                 torch.distributed.all_reduce(tmp, group=self.data_parallel_group)
                 torch.distributed.barrier()
             except (ImportError, AttributeError):
